chore: fixed code linting warnings

+++ contracts/dummy/DummyWETH.sol
@@ -5,8 +5,8 @@ import {ERC20,ERC20Burnable} from "@openzeppelin/contracts/token/ERC20/extension
 
 // dummy contract for RA and PA
 contract DummyWETH is ERC20Burnable {
-    event Deposit(address indexed dst, uint wad);
-    event Withdrawal(address indexed src, uint wad);
+    event Deposit(address indexed dst, uint256 wad);
+    event Withdrawal(address indexed src, uint256 wad);
 
     constructor() ERC20("Wrapped ETH", "WETH") {}
 
@@ -23,7 +23,7 @@ contract DummyWETH is ERC20Burnable {
         emit Deposit(msg.sender, msg.value);
     }
 
-    function withdraw(uint wad) public {
+    function withdraw(uint256 wad) public {
         _burn(msg.sender, wad);
 
         payable(msg.sender).transfer(wad);
